Add files via upload
This represents the method towards calculating machine algorithms through the ranger package, for Random-forest regression trees, in modelling occupancy of the Red-headed Woodpecker across BCR regions in North America with Landcover for environmental covariates combined with elevation and slope. 

Calibration models were also used for determination of data-inference, and to improve the accuracy of prediction models.